Record and record pattern

+++ Chapter04/P88_RecordAndInterfaces/src/main/java/modern/challenge/MelonRecord.java
@@ -1,10 +1,11 @@
 package modern.challenge;
 
-public record MelonRecord(String type, int weight) implements PestInspector {
-       
+public record MelonRecord(String type, float weight) implements PestInspector {
+
     @Override
-    public void exterminatePest() {        
-        if(detectPest()) {
+    public void exterminatePest() {
+        
+        if (detectPest()) {
             System.out.println("All pests have been exterminated");
         } else {
             System.out.println("This melon is clean, no pests have been found");
